fixed the sidebar nav that wasn't playing well with materialize and handlebars
