+++ views/layouts/main.handlebars
@@ -41,12 +41,12 @@
             </nav>
             <ul class="sidenav" id="mobile-demo">
                 {{#if loggedIn}}
-                <li><a href="/landfill/">Landfill</a></li>
-                <li><a href="/collection/">Collection</a></li>
-                <li><a href="/logout/">Logout</a></li>
+                <li class="sn"></li><a href="/landfill/">Landfill</a></li>
+                <li class="sn"><a href="/collection/">Collection</a></li>
+                <li class="sn"><a href="/logout/">Logout</a></li>
                 {{else}}
-                <li><a href="/login/">Login</a></li>
-                <li><a href="/signup/">Sign Up</a></li>
+                <li class="sn"><a href="/login/">Login</a></li>
+                <li class="sn"><a href="/login/">Sign Up</a></li>
                 {{/if}}
             </ul>
         </header>
@@ -76,8 +76,8 @@
             </div>
             <div class="footer--links">
                 <p class="bottom"> <strong class="brand">Trash<span class="twoTone">Party</span></strong> by <a href="https://github.com/Q-GAP/">Q-GAP</a> | Check Out the Code:
-                    <a href="https://github.com/Q-GAP/TrashParty"><i class="fab fa-github"> Trash Source</i></a> |
-                    <a href="mailto:support@peterroto.com"><i class="far fa-envelope"> Email Us</i></a>
+                    <a href="https://github.com/Q-GAP/TrashParty"><i class="fab fa-github"> </i>Trash Source</a> |
+                    <a href="mailto:support@peterroto.com"><i class="far fa-envelope"> </i> Email Us</a>
                     <div class="copyright" align="center">
                         <script>
                             document.write('&copy;');
@@ -89,14 +89,15 @@
                 </p>
             </div>
         </footer>
-    </section>
+  
     <!-- endregion -->
     <!-- #region JS -->
     <!-- Compiled and minified JavaScript -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
- <script src="/js/index.js"></script>
 
     <!-- endregion -->
+      </section>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
+    <script src="/js/index.js"></script>
 </body>
 
 </html>
