+++ views/login.handlebars
@@ -26,7 +26,7 @@
                             </div>
                         </div>
                         <div class="row">
-                            <a href="/login" id="loginBtn" class="btn submitBtn waves-effect waves-light" type="button">Login
+                            <a href="/login" id="loginBtn" class="btn submitBtn waves-effect waves-light" type="submit">Login
                                 <i class="material-icons right">lock_open</i>
                               </a>
                         </div>
@@ -64,12 +64,12 @@
                         </div>
                         <div class="row">
                             <div class="input-field col s10">
-                                <input id="passSignUp" type="password" class="validate">
-                                <label for="passSignUp">Password</label>
+                                <input placeholder='Password' id="passSignUp" type="password" class="validate">
+                                <label class="passLabel" for="passSignUp">Password</label>
                             </div>
                         </div>
                         <div class="row">
-                            <a href="/signup" id="signupBtn" class="btn submitBtn waves-effect waves-light" type="button">Sign Up<i class="material-icons right">person_add</i></a>
+                            <a href="/signup" id="signupBtn" class="btn submitBtn waves-effect waves-light" type="submit">Sign Up<i class="material-icons right">person_add</i></a>
                         </div>
                 </div>
             </div>
@@ -78,5 +78,6 @@
     </div>
 </div>
 </section>
-
+<script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
 <script src="/js/index.js"></script>
+<script src="/js/login.js"></script>
