+++ views/splash.handlebars
@@ -1,3 +1,6 @@
-<div id="Splash">
-    <h1>This is a placeholder splash screen.</h1>
-</div>
+{{!-- <div id="splash" class="container">
+    <h1 class="white-text">Welcome to Trash<span class="twoTone">Party</span><h1>
+</div> --}}
+<script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
+<script src="/js/index.js"></script>
+<script src="/js/splash.js"></script>
